backend/requirements.txt
Flask==3.0.1
Flask-Mail==0.9.1
APScheduler==3.10.4
flask-cors==3.0.10
reportlab==3.6.12
PyJWT==2.8.0